% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptiveInterpolation.R
\name{adaptiveInterpolation}
\alias{adaptiveInterpolation}
\title{Connects LineEnd with the nearest labeled region}
\usage{
adaptiveInterpolation(
  end_points_df,
  diagonal_edges_df,
  clean_lab_df,
  lineends_cimg
)
}
\arguments{
\item{end_points_df}{Data of type 'data.frame'.}

\item{diagonal_edges_df}{Data of type 'data.frame'.}

\item{clean_lab_df}{Data of type 'data.frame'.}

\item{lineends_cimg}{Image with dimensions of the image with discontinuous
edges. Just for giving the dimensions of the output matrix.}
}
\value{
Binary matrix that can be applied as overlay for example with
\code{\link[imager]{imager.combine}} to fill the gaps between line ends.
}
\description{
Designed to be part of the fillLineGaps function. Function scans an increasing
radius around a line end and connects it with the nearest labeled region,
which in the fillLineGaps function is the closest edge.
}
\details{
Additional details...
}
\examples{
# example code

}
