% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beads.R
\docType{data}
\name{beads}
\alias{beads}
\title{Image of microbeads}
\format{
The image was imported using imager and is therefore of class:
"cimg" "imager_array" "numeric"
}
\usage{
beads
}
\description{
This fluorescence image, formatted as 'cimg' with dimensions of
117 x 138 pixels, shows microbeads. With a single color channel, the image
provides an ideal example for in-depth analysis of microbead structures.
}
\details{
Dimensions: width - 117; height - 138; depth - 1; channel - 1
}
\examples{
data(beads)
plot(beads)
}
\references{
The image was provided by Coline Kieffer.
}
\keyword{datasets}
