% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beads_large1.R
\docType{data}
\name{beads_large1}
\alias{beads_large1}
\title{Image of microbeads}
\format{
The image was imported using imager and is therefore of class:
"cimg" "imager_array" "numeric"
}
\usage{
beads_large1
}
\description{
This fluorescence image, formatted as 'cimg' with dimensions of
492 x 376 pixels, shows microbeads. With a single color channel, the image
provides an ideal example for in-depth analysis of microbead structures.
The image's larger size encompasses a greater number of microbeads,
offering a broader range of experimental outcomes for examination.
}
\details{
Dimensions: width - 492; height - 376; depth - 1; channel - 1
}
\examples{
data(beads_large1)
plot(beads_large1)
}
\references{
The image was provided by Coline Kieffer.
}
\keyword{datasets}
