% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beads_large2.R
\docType{data}
\name{beads_large2}
\alias{beads_large2}
\title{Image of microbeads}
\format{
The image was imported using imager and is therefore of class:
"cimg" "imager_array" "numeric"
}
\usage{
beads_large2
}
\description{
This fluorescence image, formatted as 'cimg' with dimensions of
1384 x 1032 pixels, shows microbeads. With a single color channel, the image
provides an ideal example for in-depth analysis of microbead structures.
The image's larger size encompasses a greater number of microbeads,
offering a broader range of experimental outcomes for examination.
}
\details{
Dimensions: width - 1384; height - 1032; depth - 1; channel - 3
}
\examples{
data(beads_large2)
plot(beads_large2)
}
\references{
The image was provided by Coline Kieffer.
}
\keyword{datasets}
