% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detecteR.R
\name{detecteR}
\alias{detecteR}
\title{Bead Detection}
\usage{
detecteR(image, alpha = 0.75, sigma = 0.1)
}
\arguments{
\item{image}{image}

\item{alpha}{threshold adjustment factor}

\item{simge}{smoothing}
}
\value{
list of 3 objects:
\enumerate{
\item data frame of labeled region with the central coordinates
\item all coordinates that are in labeled regions
\item original image
}
}
\description{
Detects centers of beads
}
\examples{
img <- load.image(fehlt)
detecteR(img, alpha = 0.75, sigma = 0.1)
}
