% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceR.R
\name{distanceR}
\alias{distanceR}
\title{Distance based exclusion}
\usage{
distanceR(res_detecteR, radius = 10)
}
\arguments{
\item{res_detecteR}{list obtained by the detecteR function}

\item{radius}{distance in pixels from one center in which no other centers
are allowed}
}
\value{
list of 4 objects:
\enumerate{
\item data frame of labeled region with the central coordinates
\item all coordinates that are in labeled regions
\item center coordinates and cluster number of discarded regions
\item original image
}
}
\description{
Detects beads that are too close to each other
}
\examples{
res_detecteR <- detecteR(beads, alpha = 0.75, sigma = 0.1)
distanceR(res_detecteR, radius = 10)
}
