% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplet_beads.R
\docType{data}
\name{droplet_beads}
\alias{droplet_beads}
\title{Image of microbeads in luminescence channel}
\format{
The image was imported using imager and is therefore of class:
"cimg" "imager_array" "numeric"
}
\usage{
droplet_beads
}
\description{
The image shows red fluorescence rhodamine microbeads measuring
151 x 112 pixels. The fluorescence channel was used to obtain the image,
resulting in identical dimensions and positions of the beads as in the
original image (droplets).
}
\details{
Dimensions: width - 151; height - 112; depth - 1; channel - 3
}
\examples{
data(droplet_beads)
plot(droplet_beads)
}
\references{
The image was provided by Coline Kieffer.
}
\keyword{datasets}
