% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droplets.R
\docType{data}
\name{droplets}
\alias{droplets}
\title{Droplets containing microbeads}
\format{
The image was imported using imager and is therefore of class:
"cimg" "imager_array" "numeric"
}
\usage{
droplets
}
\description{
The image displays a water-oil emulsion with droplets observed through
brightfield microscopy. It is formatted as 'cimg' and sized at 151 × 112
pixels. The droplets vary in size, and some contain microbeads, which adds
complexity. Brightfield microscopy enhances the contrast between water and
oil, revealing the droplet arrangement.
}
\details{
Dimensions: width - 151; height - 112; depth - 1; channel - 1
}
\examples{
data(droplets)
plot(droplets)
}
\references{
The image was provided by Coline Kieffer.
}
\keyword{datasets}
