% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haralickCluster.R
\name{haralickCluster}
\alias{haralickCluster}
\title{k-medoids clustering of images according to the Haralick features}
\usage{
haralickCluster(path)
}
\arguments{
\item{path}{directory path to folder with images to be analyzed}
}
\value{
Data frame containing file names, md5sums and cluster number.
}
\description{
This function performs k-medoids clustering on images using Haralick
features, which describe texture. By evaluating contrast, correlation,
entropy, and homogeneity, it groups images into clusters with similar
textures. K-medoids is chosen for its outlier resilience, using actual
images as cluster centers. This approach simplifies texture-based image
analysis and classification.
}
\examples{
tempdir()
temp_dir <- tempdir()
file_path <- file.path(temp_dir, "beads.png")
save.image(beads, file_path)
file_path <- file.path(temp_dir, "droplet_beads.png")
save.image(grayscale(droplet_beads), file_path)
file_path <- file.path(temp_dir, "beads_large1.png")
save.image(beads_large1, file_path)
file_path <- file.path(temp_dir, "beads_large2.png")
save.image(grayscale(beads_large2), file_path)
result <- haralickCluster(temp_dir)
result
unlink(temp_dir, recursive = TRUE)
}
\references{
https://cran.r-project.org/package=radiomics
}
