% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolatePixels.R
\name{interpolatePixels}
\alias{interpolatePixels}
\title{Pixel Interpolation}
\usage{
interpolatePixels(row1, col1, row2, col2)
}
\arguments{
\item{row1}{first row: together with col1 coordinate for the first point}

\item{col1}{first column: together with row1 coordinate for the first point}

\item{row2}{second row: together with col2 coordinate for the second point}

\item{col2}{second column: together with row2 coordinate for the second point}
}
\value{
matrix containing the coordinates to connect the two input points
}
\description{
Connects two points in a matrix, array, or an image.
}
\examples{
test <- matrix(0, 4, 4)
test[1, 1] <- 1
test[3, 4] <- 1
link <- interpolatePixels(1, 1, 3, 4)
test[link] <- 1
}
