% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResultAnalytics.R
\name{ResultAnalytics}
\alias{ResultAnalytics}
\title{Bead Image Summary}
\usage{
ResultAnalytics(res_sizeFilter)
}
\arguments{
\item{res_sizeFilter}{list obtained by the sizeFilter function}
}
\value{
list of 2 objects:
\enumerate{
\item summary of all the microbeads in the image
\item detailed information about every single bead
}
}
\description{
Extracts all important information of the remaining microbeads. This function
summarizes the data obtained by previous functions: objectDetection,
proximityFilter and sizeFilter. Provides information like amount, intensity,
size and density.
}
\examples{
res_objectDetection <- objectDetection(beads, alpha = 0.75, sigma = 0.1)
res_proximityFilter <- proximityFilter(res_objectDetection, radius = 10)
res_sizeFilter <- sizeFilter(res_proximityFilter, lowerlimit = 50, upperlimit = 150)
ResultAnalytics(res_sizeFilter)
}
