% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeFilter.R
\name{sizeFilter}
\alias{sizeFilter}
\title{Size exclusion}
\usage{
sizeFilter(res_proximityFilter, lowerlimit = 50, upperlimit = 150)
}
\arguments{
\item{res_proximityFilter}{list obtained by the proximityFilter function}

\item{lowerlimit}{smallest accepted bead size}

\item{upperlimit}{highest accepted bead size}
}
\value{
list of 4 objects:
\enumerate{
\item remaining beads after discarding according to distance
\item all coordinates that are in labeled regions
\item size of a labeled region (bead)
\item original image
}
}
\description{
Calculates the size of a bead (in pixels) and discards too big (e.g. duplets)
or too small once
}
\examples{
res_objectDetection <- objectDetection(beads, alpha = 0.75, sigma = 0.1)
res_proximityFilter <- proximityFilter(res_objectDetection, radius = 10)
sizeFilter(res_proximityFilter, lowerlimit = 50, upperlimit = 150)
}
