% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeR.R
\name{sizeR}
\alias{sizeR}
\title{Size exclusion}
\usage{
sizeR(res_distanceR, lowerlimit = 50, upperlimit = 150)
}
\arguments{
\item{res_distanceR}{list obtained by the distanceR function}

\item{lowerlimit}{smallest accepted bead size}

\item{upperlimit}{highest accepted bead size}
}
\value{
list of 4 objects:
\enumerate{
\item remaining beads after discarding according to distance
\item all coordinates that are in labeled regions
\item size of a labeled region (bead)
\item original image
}
}
\description{
Calculates the size of a bead (in pixels) and discards too big (e.g. duplets)
or too small once
}
\examples{
img <- load.image(fehlt)
res_detecteR <- detecteR(img)
res_ distanceR <- distanceR(res_detecteR, radius = 10)
res_sizeR <- sizeR(res_distanceR, lowerlimit = 50, upperlimit = 150)
}
