% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_detectoR.R
\name{vis_detectoR}
\alias{vis_detectoR}
\title{Visualization of Bead Detection}
\usage{
vis_detectoR(image, alpha = 0.75, sigma = 0.1)
}
\arguments{
\item{image}{image}

\item{alpha}{threshold adjustment factor}

\item{sigma}{smoothing}
}
\value{
list of 1 objects:
\enumerate{
\item image were beads are marked by a purple circle
}
}
\description{
Visualizes the centers of detected beads
}
\examples{
vis_detectoR(beads, alpha = 0.75, sigma = 0.1)
}
