% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_linkeR.R
\name{vis_linkeR}
\alias{vis_linkeR}
\title{Visualization for the linkeR function}
\usage{
vis_linkeR(image, coords, color = "green")
}
\arguments{
\item{image}{image}

\item{color}{Color with which to replace specified pixels. Can be either a
an RGB triplet or one of the colors listed by \code{\link[grDevices]{colors}}.}

\item{coord}{Coordinates specifying which pixels to be colored (should
be a X|Y Data frame)}
}
\value{
plot displaying the image with highlighted pixels at desired coordinates
}
\description{
Can be generally used to change color of specified pixels in an image. In
this Package it is meant to visualize and highlight the pixels that where
added by the linkeR function.
}
\examples{
to do
}
\references{
https://CRAN.R-project.org/package=countcolors
}
