% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeR.R
\name{visualizeR}
\alias{visualizeR}
\title{Bead Image Summary}
\usage{
visualizeR(res_sizeR)
}
\arguments{
\item{res_sizeR}{list obtained by the sizeR function}
}
\value{
list of 2 objects:
\enumerate{
\item summary of whole all beads in the image
\item detailed information about every single bead
}
}
\description{
Extracts all important information of the remaining beads
}
\examples{
#' img <- load.image(fehlt)
res_detecteR <- detecteR(img)
res_ distanceR <- distanceR(res_detecteR, radius = 10)
res_sizeR <- sizeR(res_distanceR, lowerlimit = 50, upperlimit = 150)
visualizeR(res_sizeR)
}
