% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changePixelColor.R
\name{changePixelColor}
\alias{changePixelColor}
\title{Change the color of pixels}
\usage{
changePixelColor(image, coords, color = "green")
}
\arguments{
\item{image}{image}

\item{color}{Color with which to replace specified pixels. Can be either a
an RGB triplet or one of the colors listed by \code{\link[grDevices]{colors}}.}

\item{coord}{Coordinates specifying which pixels to be colored (should
be a X|Y Data frame)}
}
\value{
plot displaying the image with highlighted pixels at desired coordinates
}
\description{
Can be used to change color of specified pixels in an image. The coordinates
of the pixels must be known.
}
\examples{
coordinates <- objectDetection(test_img)
changePixelColor(test_img, coordinate$coordinates)
}
\references{
https://CRAN.R-project.org/package=countcolors
}
