% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillLineGaps.R
\name{fillLineGaps}
\alias{fillLineGaps}
\title{Reconnecting discontinuous lines}
\usage{
fillLineGaps(
  droplet.img,
  bead.img = NULL,
  threshold = "13\%",
  alpha = 0.75,
  sigma = 0.1,
  visualize = TRUE
)
}
\arguments{
\item{droplet.img}{Image that contains discontinuous lines like edges or
contours.}

\item{bead.img}{Image that contains objects that should be removed before
before applying the fill algorithm.}

\item{threshold}{"in \%" (from \code{\link[imager]{threshold}})}

\item{alpha}{Threshold adjustment factor for edge detection
(from \code{\link[imager]{cannyEdges}}).}

\item{sigma}{smoothing (from \code{\link[imager]{cannyEdges}}).}

\item{visualize}{If TRUE (default) a plot is displayed highlighting the
added pixels in the original image.}
}
\value{
image with continuous edges (closed gaps)
}
\description{
The function tries to fill in edge discontinuity so that normal labeling and
edge detection becomes possible.
}
\details{
Additional details...
}
\examples{
# example code

}
