% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_objectDetection.R
\name{interactive_objectDetection}
\alias{interactive_objectDetection}
\title{Interactive detecting of beads}
\usage{
interactive_objectDetection(image, resolution = 0.1, return_param = FALSE)
}
\arguments{
\item{image}{image (preferred import: \code{\link[imager]{load.image}})}

\item{resolution}{description}

\item{return_param}{Should the final variables for alpha and sigma should be
printed in the console (TRUE or FALSE).}
}
\description{
This function uses the objectDetection function to visualize the detected beads
at varying threshold an smoothing parameters.
}
\examples{
interactive_objectDetection(beads)
}
\references{
https://CRAN.R-project.org/package=magickGUI
}
