% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_objectDetection.R
\name{interactive_objectDetection}
\alias{interactive_objectDetection}
\title{Interactive object detection}
\usage{
interactive_objectDetection(img, resolution = 0.1, return_param = FALSE)
}
\arguments{
\item{img}{image (preferred import: \code{\link[imager]{load.image}})}

\item{resolution}{resolution of slider}

\item{return_param}{used to define the final parameter values for alpha and
sigma printed in the console (TRUE or FALSE).}
}
\value{
values of alpha and sigma
}
\description{
This function uses the objectDetection function to visualize the detected
objects at varying threshold an smoothing parameters.
}
\examples{
\donttest{
if (interactive()) {
  interactive_objectDetection(beads)
}
}
}
\references{
https://CRAN.R-project.org/package=magickGUI
}
