% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectDetection.R
\name{objectDetection}
\alias{objectDetection}
\title{Object detection}
\usage{
objectDetection(image, alpha = 0.75, sigma = 0.1)
}
\arguments{
\item{image}{image (import by \code{\link[imager]{load.image}})}

\item{alpha}{threshold adjustment factor
(from \code{\link[imager]{cannyEdges}})}

\item{sigma}{smoothing (from \code{\link[imager]{cannyEdges}})}
}
\value{
list of 4 objects:
\enumerate{
\item data frame of labeled region with the central coordinates
\item all coordinates that are in labeled regions
\item original image
\item image were beads are marked by a purple circle
}
}
\description{
This function identifies objects in an image using edge detection and
labeling, gathering the coordinates and centers of the identified objects.
Detected objects are then highlighted with colored circles for easy
recognition.
}
\examples{
objectDetection(beads, alpha = 0.75, sigma = 0.1)
}
