% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectDetection.R
\name{objectDetection}
\alias{objectDetection}
\title{Bead Detection}
\usage{
objectDetection(image, alpha = 0.75, sigma = 0.1)
}
\arguments{
\item{image}{image}

\item{alpha}{threshold adjustment factor
(from \code{\link[imager]{cannyEdges}})}

\item{sigma}{smoothing (from \code{\link[imager]{cannyEdges}})}
}
\value{
list of 3 objects:
\enumerate{
\item data frame of labeled region with the central coordinates
\item all coordinates that are in labeled regions
\item original image
\item image were beads are marked by a purple circle
}
}
\description{
Detects centers of beads
}
\examples{
objectDetection(beads, alpha = 0.75, sigma = 0.1)
}
