% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximityFilter.R
\name{proximityFilter}
\alias{proximityFilter}
\title{Distance based exclusion}
\usage{
proximityFilter(res_objectDetection, radius = 10)
}
\arguments{
\item{res_objectDetection}{list obtained by the objectDetection function}

\item{radius}{distance in pixels from one center in which no other centers
are allowed}
}
\value{
list of 4 objects:
\enumerate{
\item data frame of labeled region with the central coordinates
\item all coordinates that are in labeled regions
\item center coordinates and cluster number of discarded regions
\item original image
}
}
\description{
Detects beads that are too close to each other
}
\examples{
res_objectDetection <- objectDetection(beads, alpha = 0.75, sigma = 0.1)
proximityFilter(res_objectDetection, radius = 10)
}
